Refactor application structure by moving services and controllers to dedicated directories, and implement CasaService for house creation logic.
