+++ src/app.module.ts
@@ -3,12 +3,12 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './database/user/user.entity';
 import { Casa, FotoCasa } from './database/casa/casa.entity';
 import { CasaController } from './controllers/casa.controller';
-import { UserService } from './user.service';
-import { UserController } from './user.controller';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
+import { UserService } from './services/user.service';
+import { UserController } from './controllers/user.controller';
+import { AppController } from './controllers/app.controller';
+import { AppService } from './services/app.service';
+import { AuthService } from './services/auth.service';
+import { AuthController } from './controllers/auth.controller';
 import { EmailController } from './email/email.controller';
 import { EmailService } from './email/email.service';
 import { ComentariosModule } from './comentarios/comentarios.module';
