+++ src/services/user.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { User } from './database/user/user.entity';
+import { User } from '../database/user/user.entity';
 import * as bcrypt from 'bcrypt';
 
 @Injectable()
